Link input and output of composeUp and composeDown tasks from isRequiredBy task
Adds flag for linking input and output of composeUp and composeDown
tasks from isRequiredBy task. This allows incremental builds and avoid
the compose up and down when main task is up to date.

The flag is called `inheritInputAndOutputs` and by default is set to
false. This flag works only when isRequiredBy method is called.

This avoids declarations such as these:
composeUp {
    inputs.files test.inputs.files
    outputs.upToDateWhen {
        test
    }
}

composeDown {
    inputs.files test.inputs.files
    outputs.upToDateWhen {
        test
    }
}