feat: docker-compose v2 support (#339)
* Update configuration fixtures to version 2

* Cater for ps on non-existing containers

docker-compose v2 returns an exit code of 1 when running ps on a
non-existent container, when v1 returned 0 and an empty list.

* Remove failing container test

docker-compose v2 starts containers that depend on failed containers.
This may be considered a bug or a feature.  Either way, the only way to
recover is to bring all containers down.

* Cope with missing version in configuration

docker-compose config doesn't always include the version number from
the original configuration, so this can't be reliably used to know
whether the services are under the services key or directly under the
root.

* Cope with new container naming convention

docker-compose v2 now names containers in the form:

${container_id}_${name}-\d+

Previously, it used an underscore after the name.

* More graceful handling of missing services

docker-compose ps returns an exit code of 1 when a named service doesn't
exist.  Rather than trying to work out why the exit code is 1, get the
service list first to see whether it's worth running ps with a service
name.

* Explain reasoning behind matching logic

* More thoroughly test output capturing

Test that output from containers is captured by docker-compose.  v2
strangely strips newlines in certain circumstances.